backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
celery[redis]>=5.4.0
redis>=5.2.0
faster-whisper>=1.1.0
pyannote.audio>=3.3.0
google-genai>=1.0.0
supabase>=2.11.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
aiofiles>=24.1.0
torch>=2.5.0
torchaudio>=2.5.0
soundfile>=0.12.1
